feat: implemented deepClone function
All tests pass (I used jest) I didn't check code of tests so maybe
more testing needs to be done.

WeakMap, WeakSet and function are ignored.

I don't know the type system that is written in the comment (maybe
jshint), so my functions don't have type comments. They are simple enough so I
think it is fine as is.